refactor(cli): the binary is dispatch only; the authorization engine moves to the library (#54)
Refs #54. The 1812-line symbiote.rs becomes an entry point plus dispatch: args owns the option/flag model and the per-command applicability table, commands the operation table and help text, output the envelopes and the confirmation prompt, gate the decision to send a dangerous operation, and session the request identity and socket round trip. The policy engine moves out of the binary entirely, to the library's cli_authorization, beside the schema that publishes the very document it enforces. A move, not a redesign: all 24 probed invocations, the 4637-byte help text, plain schema stdout and schema --write bytes are unchanged, the committed fixtures are untouched, and the suite is still 114/114 for symbiote-host and 516/516 workspace-wide.
