--- crates/symbiote-host/src/bin/symbiote.rs
@@ -1,1812 +1,0 @@
-//! `symbiote` administrative CLI: a scriptable, human-readable client for a
-//! running Host daemon, using the same authenticated Unix-socket protocol as
-//! every other controller. Administrative scope only — the interactive and
-//! headless coding-agent experience (#467) is a separate surface that shares
-//! this client plumbing; no model loop exists here.
-//!
-//! Commands map one-to-one onto typed protocol operations. Every response is
-//! the daemon's own JSON (machine-readable, stable field names); `--json`
-//! wraps it in a versioned envelope for automation. Exit codes:
-//! 0 success, 1 connection/usage failure, 2 the daemon refused the command
-//! (`Err` response), 3 authorization required for a dangerous operation —
-//! distinct exits so scripts can branch on refusal vs transport failure vs
-//! a missing authorization.
-//!
-//! The CLI holds no daemon authority of its own: it cannot widen what the
-//! protocol permits. What it does own is the decision to *send* a dangerous
-//! request at all, and that decision never happens silently. An invocation
-//! authorizes a dangerous command one of three ways — `--yes`, an `yes` typed
-//! at the prompt when stdin is a terminal, or a configured **authorization
-//! policy** (`--policy FILE`, else `SYMBIOTE_CLI_POLICY`) that pre-authorizes
-//! the named operation kinds. Otherwise the CLI refuses **before connecting**,
-//! so an unattended script cannot stop a Host, start work, or smuggle either
-//! through `raw` by omitting a flag.
-//!
-//! A policy is a pre-authorization, never a widening: it is only consulted
-//! for an operation that is already dangerous, it is read with the same
-//! private-file discipline the Host applies to its own operator config
-//! (regular file, owned by this user, no group/other bits, bounded), and a
-//! policy that is missing, malformed, insecure or naming an unknown or
-//! non-dangerous kind is a usage failure (exit 1) that sends nothing. It
-//! cannot grant a permission the daemon would refuse, and it does not raise
-//! the CLI's own authority by one bit.
-use std::collections::BTreeSet;
-use std::io::{IsTerminal, Read, Write};
-use std::path::{Path, PathBuf};
-use std::process::exit;
-use std::sync::atomic::{AtomicU64, Ordering};
-
-use symbiote_host::cli_schema::{
-    CLI_SCHEMA, POLICY_SCHEMA, Risk, dangerous_kinds, document_text, known_risk_of_kind,
-    published_schemas, risk_of_kind, schema_drift, selected_schemas, selectors, write_schemas,
-};
-
-/// Exit codes, stable for scripts.
-const EXIT_OK: i32 = 0;
-const EXIT_USAGE: i32 = 1;
-const EXIT_REFUSED: i32 = 2;
-const EXIT_AUTHORIZATION_REQUIRED: i32 = 3;
-
-/// The environment variable consulted when `--policy` is absent. An empty
-/// value means "no policy", like an unset one.
-pub const POLICY_ENV: &str = "SYMBIOTE_CLI_POLICY";
-
-/// A policy grants authority, so it is held to the same bound the Host holds
-/// its own operator config to rather than being read whole.
-const POLICY_LIMIT: u64 = 64 * 1024;
-
-/// The kind the help text's policy example names. A test pins that it is still
-/// a dangerous kind, so the example can never drift into one a policy would
-/// have to reject.
-const POLICY_EXAMPLE_KIND: &str = "shutdown";
-
-fn main() {
-    match run() {
-        Ok(code) => exit(code),
-        Err(error) => {
-            eprintln!("symbiote: {error}");
-            exit(EXIT_USAGE);
-        }
-    }
-}
-
-#[derive(Debug)]
-struct Usage(String);
-
-impl std::fmt::Display for Usage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&self.0)
-    }
-}
-impl std::error::Error for Usage {}
-
-type Args<'a> = &'a [String];
-
-/// How a request came to be authorized. Kept as a value so the decision is
-/// pure and every branch is unit tested without a daemon or a terminal.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum Authorization {
-    /// The operation is not dangerous: nothing needed authorizing.
-    NotRequired,
-    /// `--yes` authorized this invocation.
-    Flag,
-    /// A configured policy pre-authorizes this operation kind.
-    Policy,
-    /// Ask on the terminal; only an explicit `yes` proceeds.
-    Prompt,
-    /// Refuse before connecting.
-    Refused,
-}
-
-/// `--yes` outranks a policy (it is the narrower, per-invocation grant) and
-/// both outrank the prompt. A policy is only ever consulted for an operation
-/// that is already dangerous, so it can never make a read or a mutation
-/// "allowed" — those never needed authorization to begin with.
-fn authorize(risk: Risk, flag: bool, policy_grant: bool, interactive: bool) -> Authorization {
-    if !risk.requires_authorization() {
-        return Authorization::NotRequired;
-    }
-    if flag {
-        return Authorization::Flag;
-    }
-    if policy_grant {
-        return Authorization::Policy;
-    }
-    if interactive {
-        Authorization::Prompt
-    } else {
-        Authorization::Refused
-    }
-}
-
-/// Only the literal `yes` (trimmed, case-insensitive) authorizes. An empty
-/// read (closed stdin) is a refusal, never a default.
-fn confirmation_accepted(line: &str) -> bool {
-    line.trim().eq_ignore_ascii_case("yes")
-}
-
-/// The parsed authorization policy: the operation kinds an operator has
-/// pre-authorized for noninteractive runs, and the point after which that
-/// pre-authorization lapses.
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct Policy {
-    authorize: BTreeSet<String>,
-    expires_at: Option<u64>,
-}
-
-impl Policy {
-    /// A policy grants exactly the dangerous kinds it lists, and only until
-    /// it expires. An expired policy is not an error: it simply grants
-    /// nothing, so the ordinary refusal path reports it.
-    fn authorizes(&self, kind: &str, now_ms: u64) -> bool {
-        if self.expires_at.is_some_and(|expiry| now_ms > expiry) {
-            return false;
-        }
-        self.authorize.contains(kind)
-    }
-
-    fn expired_at(&self, now_ms: u64) -> Option<u64> {
-        self.expires_at.filter(|expiry| now_ms > *expiry)
-    }
-}
-
-/// The policy file as written by an operator. `deny_unknown_fields` matters:
-/// a misspelled key (`authorise`, `expires`) would otherwise be ignored and
-/// the operator would believe they had configured something they had not.
-#[derive(serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct PolicyFile {
-    schema: String,
-    authorize: Vec<String>,
-    #[serde(default)]
-    expires_at: Option<u64>,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum PolicyError {
-    /// Missing, unreadable, a directory, or larger than the bound.
-    Unreadable,
-    /// Not a regular file, or readable/writable by group or other.
-    Insecure,
-    /// Not the policy schema, or an unknown/empty/non-dangerous `authorize`.
-    Invalid,
-}
-
-impl std::fmt::Display for PolicyError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(match self {
-            PolicyError::Unreadable => "cannot read the policy as a regular file",
-            PolicyError::Insecure => {
-                "the policy must be a regular file, owned by this user, with no group/other \
-                 permission bits (chmod 600)"
-            }
-            PolicyError::Invalid => "the policy is not a valid symbiote.cli-policy/v1 document",
-        })
-    }
-}
-
-impl std::error::Error for PolicyError {}
-
-/// Loads and validates a policy with the discipline the Host applies to its
-/// own private files: a symlink, a shared file, a foreign owner, an oversized
-/// file or a schema that does not name only *known dangerous* kinds is
-/// refused rather than interpreted. A policy that cannot be honored is never
-/// silently downgraded to "no policy".
-fn load_policy(path: &Path) -> Result<Policy, PolicyError> {
-    use std::os::unix::fs::{MetadataExt, OpenOptionsExt, PermissionsExt};
-    // `O_NOFOLLOW` and classify from the *opened handle*, the way the Host's
-    // own transport opens its lock file: the file that is validated is then
-    // the file that is read, so a path swapped for a symlink between a check
-    // and a read cannot slip a different document past this gate.
-    let file = match std::fs::OpenOptions::new()
-        .read(true)
-        .custom_flags(nix::fcntl::OFlag::O_NOFOLLOW.bits())
-        .open(path)
-    {
-        Ok(file) => file,
-        // `ELOOP` means the final component is a symlink: a discipline
-        // failure, not a missing file.
-        Err(error) if error.raw_os_error() == Some(nix::libc::ELOOP) => {
-            return Err(PolicyError::Insecure);
-        }
-        Err(_) => return Err(PolicyError::Unreadable),
-    };
-    let metadata = file.metadata().map_err(|_| PolicyError::Unreadable)?;
-    if !metadata.file_type().is_file()
-        || metadata.uid() != nix::unistd::geteuid().as_raw()
-        || metadata.permissions().mode() & 0o077 != 0
-    {
-        return Err(PolicyError::Insecure);
-    }
-    let mut bytes = Vec::new();
-    (&file)
-        .take(POLICY_LIMIT + 1)
-        .read_to_end(&mut bytes)
-        .map_err(|_| PolicyError::Unreadable)?;
-    if bytes.len() as u64 > POLICY_LIMIT {
-        return Err(PolicyError::Unreadable);
-    }
-    let parsed: PolicyFile = serde_json::from_slice(&bytes).map_err(|_| PolicyError::Invalid)?;
-    if parsed.schema != POLICY_SCHEMA {
-        return Err(PolicyError::Invalid);
-    }
-    if parsed.authorize.is_empty() {
-        return Err(PolicyError::Invalid);
-    }
-    let mut authorize = BTreeSet::new();
-    for kind in parsed.authorize {
-        // Only a kind that is *known and dangerous* can be named. A read or a
-        // mutation is already ungated, so listing it would be a no-op the
-        // operator could mistake for coverage; an unknown kind is one the
-        // daemon may not even define, and this CLI refuses to pre-authorize
-        // what it cannot classify. A wildcard is deliberately not a syntax.
-        if known_risk_of_kind(&kind) != Some(Risk::Dangerous) {
-            return Err(PolicyError::Invalid);
-        }
-        authorize.insert(kind);
-    }
-    Ok(Policy {
-        authorize,
-        expires_at: parsed.expires_at,
-    })
-}
-
-/// The policy path for this invocation: an explicit `--policy` wins, else
-/// `SYMBIOTE_CLI_POLICY`. An empty variable means "no policy", like unset.
-fn resolved_policy_path(options: &Options) -> Option<PathBuf> {
-    options.policy.clone().or_else(|| {
-        std::env::var_os(POLICY_ENV)
-            .map(PathBuf::from)
-            .filter(|path| !path.as_os_str().is_empty())
-    })
-}
-
-/// One command: the operation JSON builder plus the kind it emits. `raw`
-/// declares none, because the operator's file decides what it sends.
-struct Command {
-    name: &'static str,
-    summary: &'static str,
-    usage: &'static str,
-    kind: Option<&'static str>,
-    build: fn(Args, &mut serde_json::Map<String, serde_json::Value>) -> Result<(), Usage>,
-}
-
-fn field(args: Args, index: usize, name: &str) -> Result<String, Usage> {
-    args.get(index)
-        .cloned()
-        .ok_or_else(|| Usage(format!("missing <{name}>")))
-}
-
-/// Every domain identity (TaskId, HostId, DispatchId, …) serializes as a
-/// plain JSON string on the wire (serde try_from String).
-fn id_field(
-    args: Args,
-    index: usize,
-    key: &str,
-    name: &str,
-    map: &mut serde_json::Map<String, serde_json::Value>,
-) -> Result<(), Usage> {
-    let value = field(args, index, name)?;
-    map.insert(key.to_owned(), serde_json::Value::String(value));
-    Ok(())
-}
-
-fn plain(
-    key: &str,
-    value: serde_json::Value,
-    map: &mut serde_json::Map<String, serde_json::Value>,
-) {
-    map.insert(key.to_owned(), value);
-}
-
-fn read_json(path: &str) -> Result<serde_json::Value, Usage> {
-    let bytes =
-        std::fs::read(path).map_err(|error| Usage(format!("cannot read {path}: {error}")))?;
-    if bytes.len() > 60_000 {
-        return Err(Usage(format!("{path} exceeds the 64 KiB request bound")));
-    }
-    serde_json::from_slice(&bytes)
-        .map_err(|error| Usage(format!("{path} is not valid JSON: {error}")))
-}
-
-fn commands() -> Vec<Command> {
-    vec![
-        Command {
-            name: "hello",
-            summary: "negotiate protocol version and capabilities",
-            usage: "hello",
-            kind: Some("hello"),
-            build: |_, map| {
-                plain("kind", serde_json::json!("hello"), map);
-                plain(
-                    "supported_versions",
-                    serde_json::json!([symbiote_protocol::CURRENT_VERSION]),
-                    map,
-                );
-                Ok(())
-            },
-        },
-        Command {
-            name: "health",
-            summary: "check the daemon is serving",
-            usage: "health",
-            kind: Some("health"),
-            build: |_, map| {
-                plain("kind", serde_json::json!("health"), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "host-pulse",
-            summary: "owner-only passive Host resource pulse",
-            usage: "host-pulse",
-            kind: Some("get_host_pulse"),
-            build: |_, map| {
-                plain("kind", serde_json::json!("get_host_pulse"), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "shutdown",
-            summary: "drain and stop the daemon",
-            usage: "shutdown",
-            kind: Some("shutdown"),
-            build: |_, map| {
-                plain("kind", serde_json::json!("shutdown"), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "get-project",
-            summary: "read a registered Project",
-            usage: "get-project <project_id>",
-            kind: Some("get_project"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("get_project"), map);
-                plain("project_id", field(args, 0, "project_id")?.into(), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "create-task",
-            summary: "create a Task from a draft (JSON file)",
-            usage: "create-task <task.json>",
-            kind: Some("create_task"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("create_task"), map);
-                plain("task", read_json(&field(args, 0, "task.json")?)?, map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "get-task",
-            summary: "read a Task's durable state",
-            usage: "get-task <project_id> <task_id>",
-            kind: Some("get_task"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("get_task"), map);
-                plain("project_id", field(args, 0, "project_id")?.into(), map);
-                id_field(args, 1, "task_id", "task_id", map)?;
-                Ok(())
-            },
-        },
-        Command {
-            name: "read-journal",
-            summary: "read a Project's journaled events after a sequence",
-            usage: "read-journal <project_id> <after> <limit>",
-            kind: Some("read_journal"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("read_journal"), map);
-                plain("project_id", field(args, 0, "project_id")?.into(), map);
-                plain(
-                    "after",
-                    field(args, 1, "after")?
-                        .parse::<u64>()
-                        .map_err(|_| Usage("<after> must be a sequence number".into()))?
-                        .into(),
-                    map,
-                );
-                plain(
-                    "limit",
-                    field(args, 2, "limit")?
-                        .parse::<u32>()
-                        .map_err(|_| Usage("<limit> must be a page size".into()))?
-                        .into(),
-                    map,
-                );
-                Ok(())
-            },
-        },
-        Command {
-            name: "prepare-dispatch",
-            summary: "compose a dispatch preparation for a Task",
-            usage: "prepare-dispatch <task_id>",
-            kind: Some("prepare_dispatch"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("prepare_dispatch"), map);
-                id_field(args, 0, "task_id", "task_id", map)?;
-                Ok(())
-            },
-        },
-        Command {
-            name: "get-dispatch-preparation",
-            summary: "read the recorded dispatch preparation",
-            usage: "get-dispatch-preparation <task_id>",
-            kind: Some("get_dispatch_preparation"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("get_dispatch_preparation"), map);
-                id_field(args, 0, "task_id", "task_id", map)?;
-                Ok(())
-            },
-        },
-        Command {
-            name: "start-prepared-task",
-            summary: "start a prepared task on this Host",
-            usage: "start-prepared-task <task_id> <host_id>",
-            kind: Some("start_prepared_task"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("start_prepared_task"), map);
-                id_field(args, 0, "task_id", "task_id", map)?;
-                id_field(args, 1, "host_id", "host_id", map)?;
-                Ok(())
-            },
-        },
-        Command {
-            name: "run-started-dispatch",
-            summary: "activate a started dispatch (refuses without configured transports)",
-            usage: "run-started-dispatch <task_id> <dispatch_id>",
-            kind: Some("run_started_dispatch"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("run_started_dispatch"), map);
-                id_field(args, 0, "task_id", "task_id", map)?;
-                id_field(args, 1, "dispatch_id", "dispatch_id", map)?;
-                Ok(())
-            },
-        },
-        Command {
-            name: "request-task-completion",
-            summary: "file a worker completion report as evidence",
-            usage: "request-task-completion <task_id> <dispatch_id> <report>",
-            kind: Some("request_task_completion"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("request_task_completion"), map);
-                id_field(args, 0, "task_id", "task_id", map)?;
-                id_field(args, 1, "dispatch_id", "dispatch_id", map)?;
-                plain("report", field(args, 2, "report")?.into(), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "request-elevation",
-            summary: "file a worker elevation ask as evidence (never grants a lease)",
-            usage: "request-elevation <task_id> <dispatch_id> <permission> <reason>",
-            kind: Some("request_elevation"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("request_elevation"), map);
-                id_field(args, 0, "task_id", "task_id", map)?;
-                id_field(args, 1, "dispatch_id", "dispatch_id", map)?;
-                plain("permission", field(args, 2, "permission")?.into(), map);
-                plain("reason", field(args, 3, "reason")?.into(), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "scheduling-projection",
-            summary: "explain which tasks are schedulable/blocked and why",
-            usage: "scheduling-projection",
-            kind: Some("get_scheduling_projection"),
-            build: |_, map| {
-                plain("kind", serde_json::json!("get_scheduling_projection"), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "get-team",
-            summary: "read a Project's Team configuration",
-            usage: "get-team <project_id>",
-            kind: Some("get_team"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("get_team"), map);
-                plain("project_id", field(args, 0, "project_id")?.into(), map);
-                Ok(())
-            },
-        },
-        Command {
-            name: "get-task-origin",
-            summary: "read a Task's classified origin",
-            usage: "get-task-origin <project_id> <task_id>",
-            kind: Some("get_task_origin"),
-            build: |args, map| {
-                plain("kind", serde_json::json!("get_task_origin"), map);
-                plain("project_id", field(args, 0, "project_id")?.into(), map);
-                id_field(args, 1, "task_id", "task_id", map)?;
-                Ok(())
-            },
-        },
-        Command {
-            name: "raw",
-            summary: "send a raw operation JSON file (typed, validated by the daemon)",
-            usage: "raw <operation.json>",
-            // `raw` declares no kind: the authorization decision is made from
-            // the operation the file actually names (see `risk_of_kind`), so
-            // `raw` cannot be a way around the gate for `shutdown` and cannot
-            // be over-gated for a read the typed commands do not cover.
-            kind: None,
-            build: |args, map| {
-                let value = read_json(&field(args, 0, "operation.json")?)?;
-                let object = value
-                    .as_object()
-                    .ok_or_else(|| Usage("operation file must be a JSON object".into()))?;
-                for (key, value) in object {
-                    map.insert(key.clone(), value.clone());
-                }
-                Ok(())
-            },
-        },
-    ]
-}
-
-fn print_help() {
-    println!("symbiote — administrative client for a running symbioted Host");
-    println!();
-    println!(
-        "usage: symbiote [--state-dir DIR] [--command-id ID] [--policy FILE] [--json] [--yes] <command> [args...]"
-    );
-    println!(
-        "       symbiote schema [envelope|policy] [--write DIR | --check DIR]\n       symbiote help   (--command-id overrides the minted id; same id + same\n                        intent replays a lost response instead of re-executing)"
-    );
-    println!();
-    println!("options:");
-    println!("  --state-dir DIR   the private directory symbioted runs with");
-    println!("  --command-id ID   idempotency key for a retried command");
-    println!("  --policy FILE     pre-authorize dangerous operation kinds for noninteractive runs");
-    println!("  --write DIR       with `schema`, regenerate the selection into DIR");
-    println!("  --check DIR       with `schema`, compare DIR against the published documents");
-    println!("  --help, -h        print this table; connects to nothing, honors no other flag");
-    println!("  --json            one machine-readable envelope per invocation on stdout");
-    println!("  --yes             explicit authorization for this one dangerous command");
-    println!();
-    println!("flags are per command: daemon commands honor --state-dir, --command-id,");
-    println!("--policy, --json and --yes; `schema` honors --write and --check; `help` honors");
-    println!("nothing. `--help`/`-h` is universal, and a help request honors no other flag:");
-    println!("`--json --help health` is refused exactly as `--json help` is. A flag a command");
-    println!("cannot honor is a usage error that names it, never silently dropped.");
-    println!();
-    println!("responses are the daemon's JSON (pretty-printed). Exit codes:");
-    println!("0 success, 1 usage/connection failure, 2 daemon-refused command,");
-    println!("3 authorization required (no request was sent).");
-    println!();
-    println!("commands marked * can send a dangerous operation — one that starts or ends");
-    println!("execution, or grants capability:");
-    let dangerous = dangerous_kinds();
-    println!("  {}", dangerous.join(", "));
-    println!("Those are authorized by --yes, or by an interactive \"yes\" prompt when stdin is");
-    println!("a terminal, or by a policy file that names the operation kind. Otherwise they");
-    println!("refuse with exit 3 and send nothing. `raw` is marked because the operation it");
-    println!("names decides: it is authorized exactly like the equivalent typed command, reads");
-    println!("included.");
-    println!();
-    println!("a policy ({POLICY_SCHEMA}) names the kinds it pre-authorizes:");
-    println!(
-        r#"  {{"schema":"{POLICY_SCHEMA}","authorize":["{POLICY_EXAMPLE_KIND}"],"expires_at":4102444800000}}"#
-    );
-    println!("`--policy FILE`, else ${POLICY_ENV}, selects it. It must be a regular file,");
-    println!("owned by you, with no group/other permission bits (chmod 600). It grants only");
-    println!("the kinds it names and only until expires_at (optional, unix milliseconds); an");
-    println!("unreadable, insecure or invalid policy is a usage failure that sends nothing.");
-    println!();
-    println!("commands:");
-    let schema_name = "schema";
-    let schema_summary = "print the published symbiote.cli and cli-policy JSON Schemas";
-    println!("   {schema_name:<62} {schema_summary}");
-    for command in commands() {
-        let gated = match command.kind {
-            Some(kind) => risk_of_kind(kind).requires_authorization(),
-            // `raw` has no declared kind, so it MAY send a dangerous one.
-            None => true,
-        };
-        let marker = if gated { "*" } else { " " };
-        println!(" {marker} {:<62} {}", command.usage, command.summary);
-    }
-}
-
-/// The parsed command line, including the flags that decide authorization.
-#[derive(Debug)]
-struct Options {
-    state_dir: Option<PathBuf>,
-    command_id_override: Option<String>,
-    policy: Option<PathBuf>,
-    /// With `schema`, the directory to regenerate the fixture files into
-    /// instead of printing them.
-    write: Option<PathBuf>,
-    /// With `schema`, the directory to compare against the emitted documents;
-    /// reads only, so it is the non-mutating half of `--write`.
-    check: Option<PathBuf>,
-    /// `--help`/`-h`: print the command table and answer no daemon, whatever
-    /// command accompanied it.
-    help: bool,
-    json: bool,
-    yes: bool,
-    command: Option<String>,
-    args: Vec<String>,
-}
-
-/// Flags are recognized anywhere on the line — `symbiote shutdown --yes` must
-/// work as naturally as the `--yes`-first form. The first positional token is
-/// the command and the rest are its arguments; `--` ends flag parsing, so an
-/// argument that itself starts with `--` stays expressible. An unknown flag is
-/// a usage error, never silently treated as the command name or an argument.
-fn parse_options(arguments: &[String]) -> Result<Options, Usage> {
-    let mut options = Options {
-        state_dir: None,
-        command_id_override: None,
-        policy: None,
-        write: None,
-        check: None,
-        help: false,
-        json: false,
-        yes: false,
-        command: None,
-        args: Vec::new(),
-    };
-    let mut positional: Vec<String> = Vec::new();
-    let mut only_positional = false;
-    let mut index = 0;
-    while index < arguments.len() {
-        let token = arguments[index].clone();
-        if only_positional {
-            positional.push(token);
-            index += 1;
-            continue;
-        }
-        match token.as_str() {
-            "--" => only_positional = true,
-            "-h" | "--help" => options.help = true,
-            "--json" => options.json = true,
-            "--yes" => options.yes = true,
-            "--state-dir" | "--command-id" | "--policy" | "--write" | "--check" => {
-                index += 1;
-                let value = arguments
-                    .get(index)
-                    .cloned()
-                    .ok_or_else(|| Usage(format!("{token} needs a value")))?;
-                match token.as_str() {
-                    "--state-dir" => options.state_dir = Some(PathBuf::from(value)),
-                    "--write" => options.write = Some(PathBuf::from(value)),
-                    "--check" => options.check = Some(PathBuf::from(value)),
-                    "--command-id" => {
-                        // An empty idempotency key is refused rather than
-                        // carried: every empty-id invocation would collide with
-                        // every other as a replay, and the published envelope
-                        // schema requires `command_id` to be non-empty, so
-                        // accepting it would let the CLI print an envelope its
-                        // own contract rejects.
-                        if value.is_empty() {
-                            return Err(Usage(
-                                "--command-id must be a non-empty idempotency key".into(),
-                            ));
-                        }
-                        options.command_id_override = Some(value);
-                    }
-                    _ => options.policy = Some(PathBuf::from(value)),
-                }
-            }
-            flag if flag.starts_with("--") => {
-                return Err(Usage(format!("unknown option {flag}")));
-            }
-            _ => positional.push(token),
-        }
-        index += 1;
-    }
-    options.command = positional.first().cloned();
-    options.args = positional.into_iter().skip(1).collect();
-    Ok(options)
-}
-
-/// Which flags an invocation actually supplied. Checked as a whole so flag
-/// applicability is one decision instead of an `if` per flag.
-#[derive(Clone, Copy, Default)]
-struct Flags {
-    state_dir: bool,
-    command_id: bool,
-    policy: bool,
-    write: bool,
-    check: bool,
-    help: bool,
-    json: bool,
-    yes: bool,
-}
-
-impl Flags {
-    /// The supplied flags `honored` does not include, named as they are typed,
-    /// in documentation order.
-    fn unsuited_for(self, honored: Flags) -> Vec<&'static str> {
-        [
-            (self.state_dir && !honored.state_dir, "--state-dir"),
-            (self.command_id && !honored.command_id, "--command-id"),
-            (self.policy && !honored.policy, "--policy"),
-            (self.write && !honored.write, "--write"),
-            (self.check && !honored.check, "--check"),
-            (self.json && !honored.json, "--json"),
-            (self.yes && !honored.yes, "--yes"),
-            // Every command grants `--help`, so this never fires; it is
-            // listed so a future arm that forgot the grant is caught here
-            // rather than by a silently dropped `--help`.
-            (self.help && !honored.help, "--help"),
-        ]
-        .into_iter()
-        .filter(|(supplied, _)| *supplied)
-        .map(|(_, name)| name)
-        .collect()
-    }
-}
-
-impl Options {
-    fn supplied(&self) -> Flags {
-        Flags {
-            state_dir: self.state_dir.is_some(),
-            command_id: self.command_id_override.is_some(),
-            policy: self.policy.is_some(),
-            write: self.write.is_some(),
-            check: self.check.is_some(),
-            help: self.help,
-            json: self.json,
-            yes: self.yes,
-        }
-    }
-}
-
-/// The flags each command can honor. A flag outside this set is a usage error
-/// that names it, never a silent no-op. Every daemon command maps a request
-/// over the socket, so all five daemon-facing flags apply to each of them
-/// (`--policy` is consulted only for a dangerous operation and `--yes` only
-/// authorizes one, but both are valid flags on any daemon command). The local
-/// commands honor only what they use: `schema` publishes and compares
-/// documents, so `--write` and `--check`; `help` renders the command table, so
-/// nothing. `--help`/`-h` is the one universal flag, honored by every command
-/// (and by no command), because it is answered from the table alone.
-fn honored_flags(command: &str) -> Flags {
-    match command {
-        "schema" => Flags {
-            write: true,
-            check: true,
-            help: true,
-            ..Flags::default()
-        },
-        "help" => Flags {
-            help: true,
-            ..Flags::default()
-        },
-        _ => Flags {
-            state_dir: true,
-            command_id: true,
-            policy: true,
-            help: true,
-            json: true,
-            yes: true,
-            ..Flags::default()
-        },
-    }
-}
-
-/// The versioned success envelope: the daemon's own body, unmodified.
-fn success_envelope(
-    command: &str,
-    command_id: &str,
-    result: &serde_json::Value,
-) -> serde_json::Value {
-    serde_json::json!({
-        "schema": CLI_SCHEMA,
-        "command": command,
-        "command_id": command_id,
-        "ok": true,
-        "result": result,
-    })
-}
-
-/// The versioned failure envelope. `code` is the daemon's own error code, or
-/// one of the CLI's own: `authorization_required` (nothing was sent) and
-/// `unreachable` (the daemon could not be reached).
-fn error_envelope(command: &str, command_id: &str, code: &str, message: &str) -> serde_json::Value {
-    serde_json::json!({
-        "schema": CLI_SCHEMA,
-        "command": command,
-        "command_id": command_id,
-        "ok": false,
-        "error": { "code": code, "message": message },
-    })
-}
-
-/// Names BOTH the typed command and the operation that makes it dangerous:
-/// for `raw`, the operation is the whole story and the operator must see it.
-fn confirm(command: &Command, kind: &str) -> bool {
-    eprint!(
-        "symbiote: {} sends the dangerous operation {kind}; type \"yes\" to continue: ",
-        command.name
-    );
-    let _ = std::io::stderr().flush();
-    let mut line = String::new();
-    match std::io::stdin().read_line(&mut line) {
-        Ok(0) | Err(_) => false,
-        Ok(_) => confirmation_accepted(&line),
-    }
-}
-
-fn run() -> Result<i32, Box<dyn std::error::Error>> {
-    run_with(std::env::args().skip(1).collect())
-}
-
-/// The local `schema` command. An optional selector names one published
-/// document (`envelope`, `policy`) or, absent, both. `--write DIR` regenerates
-/// the selected files; `--check DIR` reports how the selected files diverge
-/// from what this binary emits and writes nothing, so it is the local half of
-/// the CI drift step. Neither touches the daemon.
-fn run_schema(options: &Options) -> Result<i32, Box<dyn std::error::Error>> {
-    let selector = match options.args.as_slice() {
-        [] => None,
-        [only] => Some(only.as_str()),
-        _ => {
-            eprintln!(
-                "symbiote: `schema` takes at most one of {}; try `symbiote help`",
-                selectors().join(" or ")
-            );
-            return Ok(EXIT_USAGE);
-        }
-    };
-    if options.write.is_some() && options.check.is_some() {
-        eprintln!("symbiote: `schema` accepts either --write DIR or --check DIR, not both");
-        return Ok(EXIT_USAGE);
-    }
-    let selected = match selected_schemas(selector) {
-        Ok(selected) => selected,
-        Err(error) => {
-            eprintln!("symbiote: {error}; try `symbiote help`");
-            return Ok(EXIT_USAGE);
-        }
-    };
-    if let Some(directory) = &options.check {
-        let drift = match schema_drift(directory, &selected) {
-            Ok(drift) => drift,
-            Err(error) => {
-                eprintln!("symbiote: {error}");
-                return Ok(EXIT_USAGE);
-            }
-        };
-        if drift.is_empty() {
-            return Ok(EXIT_OK);
-        }
-        for entry in &drift {
-            eprintln!("symbiote: {} {}", entry.file_name, entry.detail);
-        }
-        eprintln!(
-            "symbiote: {} is not what this binary publishes ({} difference(s)); `symbiote schema --write DIR` regenerates the documents",
-            directory.display(),
-            drift.len()
-        );
-        return Ok(EXIT_USAGE);
-    }
-    if let Some(directory) = &options.write {
-        write_schemas(directory, &selected)?;
-        return Ok(EXIT_OK);
-    }
-    match selector {
-        // The bare command prints one object keyed by schema identity.
-        None => println!("{}", serde_json::to_string_pretty(&published_schemas())?),
-        // A selector prints that document's canonical bytes: exactly what
-        // `schema --write` would put in its file.
-        Some(_) => print!("{}", document_text(selected[0])?),
-    }
-    Ok(EXIT_OK)
-}
-
-fn run_with(arguments: Vec<String>) -> Result<i32, Box<dyn std::error::Error>> {
-    let options = parse_options(&arguments)?;
-    // A help request — the `help` command, or `--help`/`-h` with or without a
-    // command alongside — prints the table and connects to nothing. It cannot
-    // honor any other flag either: help renders the table, not a request, so
-    // `--json`, `--state-dir` and the rest mean nothing to it.
-    let requested_help = options.help || options.command.as_deref() == Some("help");
-    // A flag a command cannot honor is a usage error that names it, never a
-    // silent no-op. This is the one place flag applicability is decided — for
-    // the local commands, the daemon commands and a help request alike —
-    // before any of them is answered or connects. A help request is validated
-    // against the `help` row, so `--json --help health` is refused by the same
-    // rule as `--json --help help`.
-    let honored = if requested_help {
-        honored_flags("help")
-    } else {
-        match options.command.as_deref() {
-            Some(name) => honored_flags(name),
-            None => {
-                // A bare invocation is a usage error, and the table is shown
-                // either way.
-                print_help();
-                return Ok(EXIT_USAGE);
-            }
-        }
-    };
-    let unsuited = options.supplied().unsuited_for(honored);
-    if !unsuited.is_empty() {
-        let named = if requested_help {
-            "help"
-        } else {
-            options.command.as_deref().unwrap_or("help")
-        };
-        eprintln!(
-            "symbiote: `{named}` does not accept {}; try `symbiote help`",
-            unsuited.join(", ")
-        );
-        return Ok(EXIT_USAGE);
-    }
-    if requested_help {
-        print_help();
-        return Ok(EXIT_OK);
-    }
-    let name = options
-        .command
-        .clone()
-        .expect("a non-help invocation names a command");
-    // `schema` is local: it emits and compares this binary's own contract and
-    // needs no daemon, no state directory and no authorization, so it is
-    // answered before any of that is consulted.
-    if name == "schema" {
-        return run_schema(&options);
-    }
-    let Some(command) = commands().into_iter().find(|c| c.name == name) else {
-        eprintln!("symbiote: unknown command {name}; try `symbiote help`");
-        return Ok(EXIT_USAGE);
-    };
-    let args = &options.args;
-    let mut operation = serde_json::Map::new();
-    if let Err(usage) = (command.build)(args, &mut operation) {
-        eprintln!("symbiote: {usage}");
-        eprintln!("usage: symbiote --state-dir DIR {}", command.usage);
-        return Ok(EXIT_USAGE);
-    }
-    // Authorization follows the OPERATION, not the command name: `raw` and the
-    // typed commands are classified by the same table, so `raw` can neither
-    // bypass the gate on `shutdown` nor be over-gated on a read.
-    let kind = operation
-        .get("kind")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or_default()
-        .to_owned();
-    let risk = risk_of_kind(&kind);
-    let now_ms = current_time_ms();
-    // A policy is consulted ONLY for an operation that is already dangerous
-    // and that `--yes` did not already authorize. Two consequences are
-    // deliberate: a read never touches the policy file, so a broken policy
-    // cannot brick `health`; and `--yes` is a complete authorization on its
-    // own, so a policy cannot block an explicitly authorized run.
-    let mut policy_error: Option<(PathBuf, PolicyError)> = None;
-    let mut policy_expired: Option<(PathBuf, u64)> = None;
-    let mut policy_grant = false;
-    if risk.requires_authorization() && !options.yes {
-        if let Some(path) = resolved_policy_path(&options) {
-            match load_policy(&path) {
-                Ok(policy) => {
-                    policy_grant = policy.authorizes(&kind, now_ms);
-                    // Report a lapse only when the policy actually named this
-                    // kind. An expired policy that never mentioned `kind` must
-                    // not be described as a lapsed grant *for* it, or the
-                    // refusal asserts something about the file that is not
-                    // true.
-                    if !policy_grant && policy.authorize.contains(&kind) {
-                        policy_expired = policy.expired_at(now_ms).map(|expiry| (path, expiry));
-                    }
-                }
-                Err(error) => policy_error = Some((path, error)),
-            }
-        }
-    }
-    // A policy that cannot be honored is an operator error, not a missing
-    // authorization: it is reported as such and nothing is sent. Silently
-    // treating it as "no policy" would let a typo quietly downgrade a
-    // pre-authorization the operator believes is in force.
-    if let Some((path, error)) = policy_error {
-        let message = format!("cannot use the policy {}: {error}", path.display());
-        if options.json {
-            println!(
-                "{}",
-                error_envelope(
-                    &name,
-                    &minted_command_id(&options, &name),
-                    "policy_invalid",
-                    &message
-                )
-            );
-            return Ok(EXIT_USAGE);
-        }
-        eprintln!("symbiote: {message}");
-        eprintln!("         no request was sent (exit 1)");
-        return Ok(EXIT_USAGE);
-    }
-    let decision = authorize(
-        risk,
-        options.yes,
-        policy_grant,
-        std::io::stdin().is_terminal(),
-    );
-    // Decided BEFORE anything touches the daemon: an unauthorized invocation
-    // must not even connect (pinned by the integration test's never-accepted
-    // socket).
-    match decision {
-        Authorization::NotRequired | Authorization::Flag | Authorization::Policy => {}
-        Authorization::Prompt => {
-            if !confirm(&command, &kind) {
-                if options.json {
-                    println!(
-                        "{}",
-                        error_envelope(
-                            &name,
-                            &minted_command_id(&options, &name),
-                            "authorization_required",
-                            "interactive confirmation was not given"
-                        )
-                    );
-                } else {
-                    eprintln!(
-                        "symbiote: {name} ({kind}) was not confirmed; no request was sent (exit 3)"
-                    );
-                }
-                return Ok(EXIT_AUTHORIZATION_REQUIRED);
-            }
-        }
-        Authorization::Refused => {
-            // The refusal names every remedy, and names the policy explicitly
-            // when one was configured but had lapsed — otherwise an operator
-            // whose policy expired sees only "--yes" and never learns why.
-            let message = match &policy_expired {
-                Some((path, expiry)) => format!(
-                    "the policy {} pre-authorized {kind} but expired at {expiry}",
-                    path.display()
-                ),
-                None => format!("noninteractive runs require --yes or a policy naming \"{kind}\""),
-            };
-            if options.json {
-                println!(
-                    "{}",
-                    error_envelope(
-                        &name,
-                        &minted_command_id(&options, &name),
-                        "authorization_required",
-                        &message
-                    )
-                );
-            } else {
-                eprintln!("symbiote: {name} sends the dangerous operation {kind}; {message}");
-                if policy_expired.is_none() {
-                    eprintln!("         (an interactive run prompts instead; no request was sent)");
-                }
-            }
-            return Ok(EXIT_AUTHORIZATION_REQUIRED);
-        }
-    }
-    let Some(directory) = options.state_dir.clone() else {
-        eprintln!("symbiote: missing --state-dir PRIVATE_DIRECTORY");
-        return Ok(EXIT_USAGE);
-    };
-    let command_id = minted_command_id(&options, &name);
-    let request = serde_json::json!({
-        "version": symbiote_protocol::CURRENT_VERSION,
-        "correlation_id": command_id,
-        "command_id": command_id,
-        "operation": operation,
-    });
-    let bytes = serde_json::to_vec(&request)?;
-    let response_bytes = match symbiote_host::transport::exchange(&directory, &bytes) {
-        Ok(bytes) => bytes,
-        Err(error) => {
-            let message = format!(
-                "cannot reach the daemon at {}: {error} (is symbioted running with --state-dir {}?)",
-                directory.join("host.sock").display(),
-                directory.display()
-            );
-            if options.json {
-                println!(
-                    "{}",
-                    error_envelope(&name, &command_id, "unreachable", &message)
-                );
-                return Ok(EXIT_USAGE);
-            }
-            eprintln!("symbiote: {message}");
-            return Ok(EXIT_USAGE);
-        }
-    };
-    let response: symbiote_protocol::Response = match serde_json::from_slice(&response_bytes) {
-        Ok(response) => response,
-        Err(error) => {
-            let message = format!("daemon sent an unparseable response: {error}");
-            if options.json {
-                println!(
-                    "{}",
-                    error_envelope(&name, &command_id, "unreachable", &message)
-                );
-                return Ok(EXIT_USAGE);
-            }
-            eprintln!("symbiote: {message}");
-            return Ok(EXIT_USAGE);
-        }
-    };
-    match &response.result {
-        Ok(body) => {
-            let value = serde_json::to_value(body).unwrap_or(serde_json::Value::Null);
-            if options.json {
-                println!("{}", success_envelope(&name, &command_id, &value));
-            } else {
-                println!(
-                    "{}",
-                    serde_json::to_string_pretty(body).unwrap_or_else(|_| format!("{body:?}"))
-                );
-            }
-            Ok(EXIT_OK)
-        }
-        Err(error) => {
-            if options.json {
-                let code = serde_json::to_value(error.code)
-                    .ok()
-                    .and_then(|value| value.as_str().map(str::to_owned))
-                    .unwrap_or_else(|| "internal".into());
-                println!(
-                    "{}",
-                    error_envelope(&name, &command_id, &code, &error.message)
-                );
-            } else {
-                eprintln!(
-                    "{}",
-                    serde_json::to_string_pretty(error).unwrap_or_else(|_| format!("{error:?}"))
-                );
-            }
-            Ok(EXIT_REFUSED)
-        }
-    }
-}
-
-/// Unix milliseconds, the unit a policy's `expires_at` is written in. A clock
-/// before the epoch reads as 0, which only makes an expiry more likely to have
-/// lapsed — the fail-closed direction.
-fn current_time_ms() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|elapsed| elapsed.as_millis() as u64)
-        .unwrap_or_default()
-}
-
-/// Distinguishes repeated minting inside one process. Its absence is a real
-/// hazard rather than a theoretical one: two invocations in the same
-/// millisecond would otherwise share an idempotency key, and the daemon would
-/// treat the second, different intent as a replay of the first.
-static NEXT_INVOCATION: AtomicU64 = AtomicU64::new(0);
-
-/// A caller-supplied `--command-id` restores the documented lost-response
-/// retry contract: the same id with identical intent replays the durable
-/// receipt instead of re-executing. The default mints an id that is unique
-/// per invocation — millisecond, pid, and an in-process sequence — so
-/// parallel scripts collide neither across processes nor within one.
-fn minted_command_id(options: &Options, name: &str) -> String {
-    match &options.command_id_override {
-        Some(id) => id.clone(),
-        None => format!(
-            "cli-{name}-{}-{}-{}",
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_millis())
-                .unwrap_or_default(),
-            std::process::id(),
-            NEXT_INVOCATION.fetch_add(1, Ordering::Relaxed)
-        ),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Commands whose *declared* operation is dangerous. `raw` declares no
-    /// kind, because the operator's file decides what it sends.
-    fn typed_dangerous_commands() -> Vec<&'static str> {
-        commands()
-            .iter()
-            .filter(|command| {
-                command
-                    .kind
-                    .is_some_and(|kind| risk_of_kind(kind).requires_authorization())
-            })
-            .map(|command| command.name)
-            .collect()
-    }
-
-    #[test]
-    fn authorization_follows_the_operation_not_the_command_name() {
-        let mut dangerous = typed_dangerous_commands();
-        dangerous.sort_unstable();
-        assert_eq!(
-            dangerous,
-            vec!["run-started-dispatch", "shutdown", "start-prepared-task"]
-        );
-
-        // One table classifies `raw` too: a dangerous operation is gated
-        // whichever route it takes, and a read through `raw` is not gated at
-        // all — otherwise the escape hatch would become unusable for exactly
-        // the operations the typed commands do not cover.
-        for kind in ["shutdown", "run_started_dispatch", "start_prepared_task"] {
-            assert!(risk_of_kind(kind).requires_authorization(), "{kind}");
-        }
-        for kind in ["health", "get_task", "read_journal", "register_project"] {
-            assert!(!risk_of_kind(kind).requires_authorization(), "{kind}");
-        }
-        assert_eq!(risk_of_kind("register_project"), Risk::Mutation);
-        assert_eq!(risk_of_kind("create_task"), Risk::Mutation);
-        assert_eq!(risk_of_kind("get_host_pulse"), Risk::ReadOnly);
-        // Grants are dangerous; the matching revocations narrow authority
-        // and are not.
-        assert!(risk_of_kind("decide_elevation").requires_authorization());
-        assert!(risk_of_kind("record_resource_consent").requires_authorization());
-        assert!(!risk_of_kind("revoke_elevation").requires_authorization());
-        assert!(!risk_of_kind("revoke_resource_consent").requires_authorization());
-        // A kind this table does not know cannot be proven safe: fail closed.
-        assert!(risk_of_kind("delete_everything").requires_authorization());
-        assert!(risk_of_kind("").requires_authorization());
-    }
-
-    #[test]
-    fn every_typed_command_emits_the_kind_it_declares() {
-        let arguments: Vec<String> = (0..6).map(|_| "1".to_string()).collect();
-        for command in commands() {
-            // The two file-taking commands read an operator-supplied file.
-            if matches!(command.name, "create-task" | "raw") {
-                continue;
-            }
-            let mut operation = serde_json::Map::new();
-            if let Err(error) = (command.build)(&arguments, &mut operation) {
-                panic!(
-                    "{} must build with synthetic arguments: {error}",
-                    command.name
-                );
-            }
-            let emitted = operation
-                .get("kind")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or_default();
-            assert_eq!(
-                command.kind,
-                Some(emitted),
-                "{} must declare the kind it emits",
-                command.name
-            );
-        }
-        let undeclared: Vec<&str> = commands()
-            .iter()
-            .filter(|command| command.kind.is_none())
-            .map(|command| command.name)
-            .collect();
-        assert_eq!(undeclared, vec!["raw"]);
-    }
-
-    #[test]
-    fn read_only_and_mutating_commands_never_prompt() {
-        for risk in [Risk::ReadOnly, Risk::Mutation] {
-            assert_eq!(
-                authorize(risk, false, false, false),
-                Authorization::NotRequired
-            );
-            assert_eq!(
-                authorize(risk, false, false, true),
-                Authorization::NotRequired
-            );
-            assert_eq!(
-                authorize(risk, true, false, false),
-                Authorization::NotRequired
-            );
-        }
-    }
-
-    #[test]
-    fn dangerous_commands_refuse_without_authorization_or_a_terminal() {
-        // No flag, no policy, no terminal: refuse — never a silent default.
-        assert_eq!(
-            authorize(Risk::Dangerous, false, false, false),
-            Authorization::Refused
-        );
-        // An interactive terminal prompts instead of refusing.
-        assert_eq!(
-            authorize(Risk::Dangerous, false, false, true),
-            Authorization::Prompt
-        );
-        // `--yes` is honored in both modes and is reported as the flag's grant.
-        assert_eq!(
-            authorize(Risk::Dangerous, true, false, false),
-            Authorization::Flag
-        );
-        assert_eq!(
-            authorize(Risk::Dangerous, true, false, true),
-            Authorization::Flag
-        );
-    }
-
-    #[test]
-    fn a_policy_grant_is_authorization_without_a_flag_or_a_terminal() {
-        assert_eq!(
-            authorize(Risk::Dangerous, false, true, false),
-            Authorization::Policy
-        );
-        assert_eq!(
-            authorize(Risk::Dangerous, false, true, true),
-            Authorization::Policy
-        );
-        // The per-invocation flag is the narrower grant, so it is the one
-        // reported when both apply.
-        assert_eq!(
-            authorize(Risk::Dangerous, true, true, false),
-            Authorization::Flag
-        );
-    }
-
-    #[test]
-    fn only_an_explicit_yes_confirms() {
-        assert!(confirmation_accepted("yes"));
-        assert!(confirmation_accepted("YES"));
-        assert!(confirmation_accepted("  Yes\n"));
-        // Everything else — including a closed stdin's empty read — refuses.
-        for line in ["", "\n", "y", "no", "yes please", "shutdown"] {
-            assert!(!confirmation_accepted(line), "{line:?} must not confirm");
-        }
-    }
-
-    fn options(arguments: &[&str]) -> Options {
-        parse_options(&arguments.iter().map(|s| s.to_string()).collect::<Vec<_>>()).unwrap()
-    }
-
-    #[test]
-    fn flags_parse_before_and_after_the_command() {
-        let before = options(&["--json", "--yes", "--state-dir", "/s", "shutdown"]);
-        assert!(before.json && before.yes);
-        assert_eq!(before.state_dir, Some(PathBuf::from("/s")));
-        assert_eq!(before.command.as_deref(), Some("shutdown"));
-        assert!(before.args.is_empty());
-
-        // The natural form — flags after the command — must behave the same,
-        // or `symbiote shutdown --yes` would silently refuse.
-        let after = options(&["shutdown", "--yes", "--json"]);
-        assert!(after.json && after.yes);
-        assert_eq!(after.command.as_deref(), Some("shutdown"));
-        assert!(after.args.is_empty());
-
-        let mixed = options(&[
-            "--state-dir",
-            "/s",
-            "--command-id",
-            "c1",
-            "get-task",
-            "p",
-            "t",
-        ]);
-        assert_eq!(mixed.command_id_override.as_deref(), Some("c1"));
-        assert_eq!(mixed.args, vec!["p".to_string(), "t".to_string()]);
-    }
-
-    #[test]
-    fn an_unknown_or_incomplete_flag_is_a_usage_error_not_a_command() {
-        // An unknown flag is refused wherever it appears, and a flag that
-        // needs a value is never satisfied by the command name.
-        for (arguments, named) in [
-            (vec!["--bare".to_string()], "--bare"),
-            (vec!["--state-dir".to_string()], "--state-dir"),
-            (
-                vec!["raw".to_string(), "--yes.json".to_string()],
-                "--yes.json",
-            ),
-        ] {
-            let error = parse_options(&arguments).unwrap_err();
-            assert!(error.to_string().contains(named), "{error}");
-        }
-        // `--` keeps an argument that starts with dashes expressible.
-        let escaped = options(&["raw", "--", "--yes.json"]);
-        assert_eq!(escaped.command.as_deref(), Some("raw"));
-        assert_eq!(escaped.args, vec!["--yes.json".to_string()]);
-        // ... and the value of a value-taking flag is not re-read as one.
-        let literal = options(&["--command-id", "--yes", "health"]);
-        assert_eq!(literal.command_id_override.as_deref(), Some("--yes"));
-        assert_eq!(literal.command.as_deref(), Some("health"));
-        assert!(!literal.yes);
-    }
-
-    #[test]
-    fn flag_applicability_is_declared_per_command() {
-        // One table decides which flags each command can honor, so adding a
-        // local command means naming its flags there rather than sprinkling
-        // `if` checks through the handler. `schema` honors `--write` and
-        // `--check`, `help` honors nothing, and daemon commands honor the
-        // daemon flags. `--help`/`-h` is honored everywhere.
-        for arguments in [
-            vec!["--write", "/tmp/s", "schema"],
-            vec!["--check", "/tmp/s", "schema"],
-            vec!["--write", "/tmp/s", "--check", "/tmp/s", "schema"],
-        ] {
-            let supplied = options(&arguments).supplied();
-            assert!(
-                supplied.unsuited_for(honored_flags("schema")).is_empty(),
-                "{arguments:?}"
-            );
-        }
-        for (arguments, named) in [
-            (vec!["--json", "schema"], "--json"),
-            (vec!["--yes", "help"], "--yes"),
-            (vec!["--write", "/tmp/s", "health"], "--write"),
-            (vec!["--check", "/tmp/s", "health"], "--check"),
-        ] {
-            let supplied = options(&arguments).supplied();
-            let command = arguments.last().unwrap();
-            assert_eq!(
-                supplied.unsuited_for(honored_flags(command)),
-                vec![named],
-                "{arguments:?}"
-            );
-        }
-        let daemon = options(&["--state-dir", "/s", "--json", "--yes", "shutdown"]);
-        assert!(
-            daemon
-                .supplied()
-                .unsuited_for(honored_flags("shutdown"))
-                .is_empty()
-        );
-        // `--help` is universal: no command reports it unsuited.
-        for command in ["schema", "help", "shutdown", "bogus"] {
-            assert!(
-                options(&["--help", command])
-                    .supplied()
-                    .unsuited_for(honored_flags(command))
-                    .is_empty(),
-                "{command}"
-            );
-        }
-    }
-
-    #[test]
-    fn help_is_a_universal_flag_that_connects_to_nothing() {
-        // `--help`/`-h` are flags, not commands, and are answered from the
-        // table alone: alone, or with any command, they print the table and
-        // exit 0 without a state directory or a socket. This fails against
-        // the earlier parser, which rejected `--help` as an unknown option and
-        // treated `-h` as an unknown command.
-        for arguments in [
-            vec!["--help"],
-            vec!["-h"],
-            vec!["--help", "shutdown"],
-            vec!["shutdown", "--help"],
-            vec!["schema", "--help"],
-            vec!["health", "-h"],
-        ] {
-            let code = run_with(arguments.iter().map(|s| s.to_string()).collect()).unwrap();
-            assert_eq!(code, EXIT_OK, "{arguments:?}");
-        }
-        // With no command and no `--help`, the table is still a usage error.
-        assert_eq!(run_with(Vec::new()).unwrap(), EXIT_USAGE);
-        // A help request is validated against the `help` row, so a flag it
-        // cannot honor is refused identically whether the command named honors
-        // that flag or not: `--json --help health` behaves like `--json help`
-        // rather than exiting 0 and dropping `--json`.
-        for arguments in [
-            vec!["--json", "help"],
-            vec!["--json", "--help"],
-            vec!["--json", "--help", "help"],
-            vec!["--json", "--help", "health"],
-            vec!["--help", "health", "--state-dir", "/d"],
-            vec!["--write", "/d", "schema", "--help"],
-        ] {
-            assert_eq!(
-                run_with(arguments.iter().map(|s| s.to_string()).collect()).unwrap(),
-                EXIT_USAGE,
-                "{arguments:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn envelopes_carry_the_schema_command_and_id() {
-        let result = serde_json::json!({"kind": "shutdown"});
-        let success = success_envelope("shutdown", "cli-1", &result);
-        assert_eq!(success["schema"], CLI_SCHEMA);
-        assert_eq!(success["command"], "shutdown");
-        assert_eq!(success["command_id"], "cli-1");
-        assert_eq!(success["ok"], true);
-        assert_eq!(success["result"], result);
-
-        let failure = error_envelope("shutdown", "cli-1", "authorization_required", "no");
-        assert_eq!(failure["schema"], CLI_SCHEMA);
-        assert_eq!(failure["ok"], false);
-        assert_eq!(failure["error"]["code"], "authorization_required");
-        assert!(failure.get("result").is_none());
-    }
-
-    /// The published fixtures under `docs/contracts/schemas/` are the external
-    /// contract. These assertions bind them to this binary, so a change here
-    /// cannot leave the published documents describing something else.
-    fn load_fixture(name: &str) -> serde_json::Value {
-        let path = symbiote_host::cli_schema::fixture_directory().join(name);
-        let text = std::fs::read_to_string(&path)
-            .unwrap_or_else(|error| panic!("{}: {error}", path.display()));
-        serde_json::from_str(&text)
-            .unwrap_or_else(|error| panic!("{} is not JSON: {error}", path.display()))
-    }
-
-    #[test]
-    fn the_published_schemas_track_the_cli_contract() {
-        let envelope = load_fixture("symbiote.cli.v1.schema.json");
-        assert_eq!(envelope["properties"]["schema"]["const"], CLI_SCHEMA);
-        // The declared keys are the whole shape: with `additionalProperties:
-        // false` in the fixture, an envelope may carry these and no others.
-        let declared: BTreeSet<&str> = envelope["properties"]
-            .as_object()
-            .unwrap()
-            .keys()
-            .map(String::as_str)
-            .collect();
-        assert_eq!(
-            declared,
-            BTreeSet::from(["command", "command_id", "error", "ok", "result", "schema"])
-        );
-        assert_eq!(
-            envelope["required"],
-            serde_json::json!(["schema", "command", "command_id", "ok"])
-        );
-        // The non-empty `command_id` the fixture requires is enforced by the
-        // parser, not merely documented here: see the empty-override test below.
-        assert_eq!(envelope["properties"]["command_id"]["minLength"], 1);
-        assert_eq!(envelope["oneOf"].as_array().unwrap().len(), 2);
-        assert_eq!(
-            envelope["$defs"]["error"]["required"],
-            serde_json::json!(["code", "message"])
-        );
-        // Both constructors write a subset of the declared keys, and exactly
-        // one of result/error — the exclusivity the fixture encodes as oneOf
-        // and a plain struct cannot.
-        let success = success_envelope("health", "cli-1", &serde_json::json!({"kind": "hello"}));
-        let failure = error_envelope("health", "cli-1", "unreachable", "down");
-        for written in [&success, &failure] {
-            let keys: BTreeSet<&str> = written
-                .as_object()
-                .unwrap()
-                .keys()
-                .map(String::as_str)
-                .collect();
-            assert!(
-                keys.is_subset(&declared),
-                "{keys:?} must stay within {declared:?}"
-            );
-        }
-        assert!(success.get("result").is_some() && success.get("error").is_none());
-        assert!(failure.get("error").is_some() && failure.get("result").is_none());
-
-        let policy = load_fixture("symbiote.cli-policy.v1.schema.json");
-        assert_eq!(policy["properties"]["schema"]["const"], POLICY_SCHEMA);
-        assert_eq!(
-            policy["required"],
-            serde_json::json!(["schema", "authorize"])
-        );
-        assert_eq!(policy["properties"]["authorize"]["minItems"], 1);
-        // The policy schema may name exactly the kinds this CLI classifies as
-        // dangerous: no more (it would publish a grant the CLI refuses) and no
-        // fewer (a new dangerous kind would be un-nameable by any policy).
-        let mut published: Vec<&str> = policy["properties"]["authorize"]["items"]["enum"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|kind| kind.as_str().unwrap())
-            .collect();
-        published.sort_unstable();
-        let mut dangerous = dangerous_kinds();
-        dangerous.sort_unstable();
-        assert_eq!(published, dangerous);
-        assert!(published.contains(&POLICY_EXAMPLE_KIND));
-    }
-
-    #[test]
-    fn a_minted_command_id_is_per_invocation_and_override_is_honored() {
-        let mut options = parse_options(&["health".to_string()]).unwrap();
-        let first = minted_command_id(&options, "health");
-        let second = minted_command_id(&options, "health");
-        assert!(first.starts_with("cli-health-"));
-        assert_ne!(first, second);
-        options.command_id_override = Some("retry-1".into());
-        assert_eq!(minted_command_id(&options, "health"), "retry-1");
-    }
-
-    #[test]
-    fn an_empty_command_id_override_is_refused() {
-        // The published envelope schema requires a non-empty `command_id`, so
-        // the parser must not carry an empty override through to an envelope:
-        // that would make the CLI violate its own contract, and an empty
-        // idempotency key makes every such invocation a replay of the last.
-        let empty = parse_options(&[
-            "--command-id".to_string(),
-            String::new(),
-            "health".to_string(),
-        ]);
-        assert!(matches!(empty, Err(Usage(ref message)) if message.contains("command-id")));
-        // A non-empty override is still honored, and the flag still works after
-        // the command, so the refusal is only about the empty value.
-        let after = options(&["health", "--command-id", "retry-1"]);
-        assert_eq!(after.command_id_override.as_deref(), Some("retry-1"));
-    }
-
-    const SHUTDOWN_ONLY: &str = r#"{"schema":"symbiote.cli-policy/v1","authorize":["shutdown"]}"#;
-
-    static NEXT_POLICY_FILE: AtomicU64 = AtomicU64::new(0);
-
-    /// Writes a policy file with an explicit mode and returns its path; the
-    /// test removes it. Uniqueness is per (process, call) so parallel tests
-    /// cannot see each other's file.
-    fn write_policy(contents: &str, mode: u32) -> PathBuf {
-        use std::os::unix::fs::PermissionsExt;
-        let path = std::env::temp_dir().join(format!(
-            "symbiote-cli-policy-unit-{}-{}",
-            std::process::id(),
-            NEXT_POLICY_FILE.fetch_add(1, Ordering::Relaxed)
-        ));
-        std::fs::write(&path, contents).unwrap();
-        std::fs::set_permissions(&path, std::fs::Permissions::from_mode(mode)).unwrap();
-        path
-    }
-
-    fn remove(path: &Path) {
-        std::fs::remove_file(path).unwrap();
-    }
-
-    #[test]
-    fn a_private_policy_authorizes_exactly_the_dangerous_kinds_it_names() {
-        let path = write_policy(SHUTDOWN_ONLY, 0o600);
-        let policy = load_policy(&path).expect("a 0600 policy is honored");
-        assert_eq!(policy.expires_at, None);
-        assert!(policy.authorizes("shutdown", 0));
-        // Not a wildcard: a kind the policy does not name stays unauthorized,
-        // and a kind that was never gated stays ungated rather than "granted".
-        for kind in ["start_prepared_task", "decide_elevation", "health"] {
-            assert!(!policy.authorizes(kind, 0), "{kind}");
-        }
-        remove(&path);
-    }
-
-    #[test]
-    fn a_policy_lapses_after_its_expiry_rather_than_erroring() {
-        let path = write_policy(
-            r#"{"schema":"symbiote.cli-policy/v1","authorize":["shutdown"],"expires_at":1000}"#,
-            0o600,
-        );
-        let policy = load_policy(&path).unwrap();
-        // The expiry instant itself is still authorized; the next millisecond
-        // is not. An expired policy grants nothing but is not a load failure —
-        // the ordinary refusal path reports it.
-        assert!(policy.authorizes("shutdown", 1000));
-        assert!(!policy.authorizes("shutdown", 1001));
-        assert_eq!(policy.expired_at(1000), None);
-        assert_eq!(policy.expired_at(1001), Some(1000));
-        remove(&path);
-    }
-
-    #[test]
-    fn every_dangerous_kind_in_the_table_can_be_named_by_a_policy() {
-        let dangerous = dangerous_kinds();
-        // Pins the set the help text and the policy schema describe.
-        assert_eq!(
-            dangerous,
-            vec![
-                "start_prepared_task",
-                "run_started_dispatch",
-                "shutdown",
-                "decide_elevation",
-                "record_resource_consent",
-            ]
-        );
-        let contents = format!(
-            r#"{{"schema":"symbiote.cli-policy/v1","authorize":[{}]}}"#,
-            dangerous
-                .iter()
-                .map(|kind| format!("\"{kind}\""))
-                .collect::<Vec<_>>()
-                .join(",")
-        );
-        let path = write_policy(&contents, 0o600);
-        // The help text's example must stay a kind a policy accepts.
-        assert!(dangerous.contains(&POLICY_EXAMPLE_KIND));
-        let policy = load_policy(&path).expect("every dangerous kind is nameable");
-        for kind in dangerous {
-            assert!(policy.authorizes(kind, 0), "{kind}");
-        }
-        remove(&path);
-    }
-
-    #[test]
-    fn a_policy_that_cannot_be_honored_is_refused_not_ignored() {
-        // A policy grants authority, so it is read strictly: an unknown schema
-        // version, a missing or empty list, a wildcard, a typo'd or unknown
-        // key, and — deliberately — a kind that is known but NOT dangerous
-        // (listing it would be a no-op the operator could mistake for
-        // coverage), plus a kind this build cannot classify at all.
-        for contents in [
-            r#"{"schema":"symbiote.cli-policy/v2","authorize":["shutdown"]}"#,
-            r#"{"authorize":["shutdown"]}"#,
-            r#"{"schema":"symbiote.cli-policy/v1"}"#,
-            r#"{"schema":"symbiote.cli-policy/v1","authorize":[]}"#,
-            r#"{"schema":"symbiote.cli-policy/v1","authorize":["*"]}"#,
-            r#"{"schema":"symbiote.cli-policy/v1","authorize":["shutdown","health"]}"#,
-            r#"{"schema":"symbiote.cli-policy/v1","authorize":["create_task"]}"#,
-            r#"{"schema":"symbiote.cli-policy/v1","authorize":["delete_everything"]}"#,
-            r#"{"schema":"symbiote.cli-policy/v1","authorise":["shutdown"]}"#,
-            r#"{"schema":"symbiote.cli-policy/v1","authorize":["shutdown"],"expires":"soon"}"#,
-            "not json",
-        ] {
-            let path = write_policy(contents, 0o600);
-            assert_eq!(load_policy(&path), Err(PolicyError::Invalid), "{contents}");
-            remove(&path);
-        }
-        // A policy that cannot be read at all is Unreadable, not Invalid.
-        assert_eq!(
-            load_policy(Path::new("/nonexistent/symbiote-policy.json")),
-            Err(PolicyError::Unreadable)
-        );
-    }
-
-    #[test]
-    fn a_shared_symlinked_oversized_or_non_file_policy_is_refused() {
-        use std::os::unix::fs::PermissionsExt;
-        // Any group/other bit is authority a neighbor could rewrite.
-        for mode in [0o640, 0o604, 0o606, 0o660, 0o666, 0o777] {
-            let path = write_policy(SHUTDOWN_ONLY, mode);
-            assert_eq!(load_policy(&path), Err(PolicyError::Insecure), "{mode:o}");
-            remove(&path);
-        }
-        // Owner-only 0400 and 0600 are both honored: the bound is *sharing*,
-        // not writability.
-        for mode in [0o400, 0o600] {
-            let path = write_policy(SHUTDOWN_ONLY, mode);
-            assert!(load_policy(&path).is_ok(), "{mode:o}");
-            remove(&path);
-        }
-        // A symlink is refused rather than followed: the file that grants
-        // authority must be the file the operator inspected.
-        let target = write_policy(SHUTDOWN_ONLY, 0o600);
-        let link = target.with_extension("link");
-        std::os::unix::fs::symlink(&target, &link).unwrap();
-        assert_eq!(load_policy(&link), Err(PolicyError::Insecure));
-        std::fs::remove_file(&link).unwrap();
-        remove(&target);
-        // A directory is not a policy.
-        let directory = std::env::temp_dir().join(format!(
-            "symbiote-cli-policy-dir-{}-{}",
-            std::process::id(),
-            NEXT_POLICY_FILE.fetch_add(1, Ordering::Relaxed)
-        ));
-        std::fs::create_dir_all(&directory).unwrap();
-        std::fs::set_permissions(&directory, std::fs::Permissions::from_mode(0o700)).unwrap();
-        assert_eq!(load_policy(&directory), Err(PolicyError::Insecure));
-        std::fs::remove_dir_all(&directory).unwrap();
-        // Bounded: a policy past 64 KiB is refused, and the bound is checked
-        // before the document is interpreted (the padding key would otherwise
-        // be an unknown field).
-        let oversized = format!(
-            r#"{{"schema":"symbiote.cli-policy/v1","authorize":["shutdown"],"padding":"{}"}}"#,
-            "x".repeat(64 * 1024)
-        );
-        let path = write_policy(&oversized, 0o600);
-        assert_eq!(load_policy(&path), Err(PolicyError::Unreadable));
-        remove(&path);
-    }
-
-    #[test]
-    fn the_policy_flag_selects_the_file_before_or_after_the_command() {
-        for arguments in [
-            vec!["--policy", "/p.json", "shutdown"],
-            vec!["shutdown", "--policy", "/p.json"],
-        ] {
-            let parsed = options(&arguments);
-            assert_eq!(parsed.policy, Some(PathBuf::from("/p.json")));
-            assert_eq!(parsed.command.as_deref(), Some("shutdown"));
-            assert_eq!(
-                resolved_policy_path(&parsed),
-                Some(PathBuf::from("/p.json"))
-            );
-        }
-        assert_eq!(options(&["health"]).policy, None);
-    }
-}
